Kcc Query Assistant/requirements.txt
# Core ML and NLP
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.3

# Vector Database
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Web Framework
streamlit>=1.28.0
streamlit-chat>=0.1.1

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Optional: GPU support (uncomment if you have CUDA)
# torch>=2.0.0+cu118 --index-url https://download.pytorch.org/whl/cu118
# faiss-gpu>=1.7.4